backend/requirements.txt
fastapi==0.114.0
uvicorn==0.23.1
langchain==1.2.12
langchain-community==0.3.31
ollama==0.6.1
pydantic>=2.9.3,<3.0.0
requests>=2.32.5,<3.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0